Move apache collector functions and flags into that collector file
Rename --listen flag to --web.listen-address
Rename --apache-status  to --collector.apache.status-url
